refactor(html): unify entity decoding and move parsing helpers into syntax.js
Fold decodeEntitiesWithMap into decodeEntities behind a `withMap`
option: both modes share the hand-rolled reference scanner (the map
variant previously kept its own regex-exec loop), with the map
bookkeeping outlined into helpers so the dominant no-map mode keeps its
exact hot path. Overloaded JSDoc preserves the per-mode return types;
a 60k-case differential fuzz agrees with the old pair in both modes.

Move the remaining spec-level parsing helpers out of HtmlParser into
syntax.js next to parseSrcset: parseSrc, parseMsapplicationTask,
parseCssUrls, attrSourceSpan and captureIntegrityRange. The duplicated
ASCII-whitespace predicate collapses onto the tokenizer's (exported as
isAsciiWhitespace), and HtmlParser's eager `../css/syntax` require
becomes lazy — first parseCssUrls call — matching HtmlGenerator's
deliberate deferral.
