+++ lib/html/HtmlParser.js
@@ -5,15 +5,6 @@
 "use strict";
 
 const Parser = require("../Parser");
-const {
-	TT_EOF,
-	TT_FUNCTION,
-	TT_STRING,
-	TT_URL,
-	TT_WHITESPACE,
-	TokenStream,
-	equalsLowerCase
-} = require("../css/syntax");
 const ConstDependency = require("../dependencies/ConstDependency");
 const HtmlEntryDependency = require("../dependencies/HtmlEntryDependency");
 const HtmlInlineHtmlDependency = require("../dependencies/HtmlInlineHtmlDependency");
@@ -40,10 +31,15 @@ const {
 	NodeType,
 	SVG_TAG_ADJUST,
 	SourceProcessor,
+	attrSourceSpan,
+	captureIntegrityRange,
 	decodeEntities,
-	decodeEntitiesWithMap,
 	escapeAttribute,
 	escapeText,
+	isAsciiWhitespace,
+	parseCssUrls,
+	parseMsapplicationTask,
+	parseSrc,
 	parseSrcset
 } = require("./syntax");
 
@@ -70,31 +66,6 @@ const {
  * @property {(error: Error | string) => void} emitError report an error on the module
  */
 /** @typedef {(source: string, context: HtmlTemplateContext) => string} HtmlTemplateFunction */
-
-const HORIZONTAL_TAB = "\u0009".charCodeAt(0);
-const NEWLINE = "\u000A".charCodeAt(0);
-const FORM_FEED = "\u000C".charCodeAt(0);
-const CARRIAGE_RETURN = "\u000D".charCodeAt(0);
-const SPACE = "\u0020".charCodeAt(0);
-
-/**
- * @param {number} char char
- * @returns {boolean} true when ASCII whitespace, otherwise false
- */
-function isASCIIWhitespace(char) {
-	return (
-		// Horizontal tab
-		char === HORIZONTAL_TAB ||
-		// New line
-		char === NEWLINE ||
-		// Form feed
-		char === FORM_FEED ||
-		// Carriage return
-		char === CARRIAGE_RETURN ||
-		// Space
-		char === SPACE
-	);
-}
 
 /** @typedef {import("./syntax").ParsedSource} ParsedSource */
 
@@ -180,181 +151,7 @@ const buildHeadTags = (opts) => {
 // source order.
 const COPYABLE_SIBLING_ATTRS = ["nonce", "crossorigin", "referrerpolicy"];
 
-const CC_QUOTATION = '"'.charCodeAt(0);
-const CC_APOSTROPHE = "'".charCodeAt(0);
 const CC_SLASH = "/".charCodeAt(0);
-
-/**
- * Byte-exact source span of an attribute including the single leading
- * whitespace (` name`, ` name=value`, ` name="value"`), mirroring the
- * tokenizer's end-of-attribute rule (`valueEnd + 1` past the closing quote).
- * @param {import("./syntax").HtmlPath} path walk path (used only for attribute-ref reads)
- * @param {string} source HTML source
- * @param {import("./syntax").HtmlAttributeRef} attr attribute ref
- * @returns {string} the attribute's source slice
- */
-const attrSourceSpan = (path, source, attr) => {
-	const valueStart = path.attributeValueStart(attr);
-	const valueEnd = path.attributeValueEnd(attr);
-	const end =
-		valueStart === -1
-			? path.attributeNameEnd(attr)
-			: source.charCodeAt(valueStart - 1) === CC_QUOTATION ||
-				  source.charCodeAt(valueStart - 1) === CC_APOSTROPHE
-				? valueEnd + 1
-				: valueEnd;
-	return source.slice(path.attributeNameStart(attr) - 1, end);
-};
-
-/**
- * Source span of a tag's `integrity` attribute (offsets relative to
- * `elementStart`, leading whitespace and quotes included), so the template can
- * drop the content-specific author value without re-parsing. Null when absent.
- * @param {import("./syntax").HtmlPath} path walk path (for attribute reads)
- * @param {string} source HTML source
- * @param {number} elementStart offset of the tag's opening `<`
- * @returns {Range | null} the span, or null when there is no `integrity`
- */
-const captureIntegrityRange = (path, source, elementStart) => {
-	const integrityAttr = path.findAttribute("integrity");
-	if (integrityAttr === 0 || path.attributeNameStart(integrityAttr) < 0) {
-		return null;
-	}
-	let rs = path.attributeNameStart(integrityAttr);
-	while (rs > 0 && isASCIIWhitespace(source.charCodeAt(rs - 1))) {
-		rs--;
-	}
-	let re;
-	const ivs = path.attributeValueStart(integrityAttr);
-	if (ivs === -1) {
-		re = path.attributeNameEnd(integrityAttr);
-	} else {
-		const q = source.charCodeAt(ivs - 1);
-		const ive = path.attributeValueEnd(integrityAttr);
-		re = q === CC_QUOTATION || q === CC_APOSTROPHE ? ive + 1 : ive;
-	}
-	return [rs - elementStart, re - elementStart];
-};
-
-// eslint-disable-next-line no-control-regex
-const IGNORE_CHARS_REGEXP = /[\u0000-\u001F\u007F-\u009F\u00A0]/g;
-
-/**
- * @param {string} input input
- * @returns {ParsedSource[]} parsed src
- */
-const parseSrc = (input) => {
-	const len = input.length;
-	if (len === 0) throw new Error("Must be non-empty");
-
-	let start = 0;
-	let end = len;
-
-	while (start < end) {
-		const code = input.charCodeAt(start);
-		if (code > 32 && code !== 160) break;
-		start++;
-	}
-
-	if (start === end) throw new Error("Must be non-empty");
-
-	while (end > start) {
-		const code = input.charCodeAt(end - 1);
-		if (code > 32 && code !== 160) break;
-		end--;
-	}
-
-	let value = input.slice(start, end);
-
-	if (IGNORE_CHARS_REGEXP.test(value)) {
-		value = value.replace(IGNORE_CHARS_REGEXP, "");
-		if (value.length === 0) throw new Error("Must be non-empty");
-	}
-
-	return [[value, start, end]];
-};
-
-/**
- * Extracts the `icon-uri` value of an `msapplication-task` meta content
- * (`name=…;action-uri=…;icon-uri=…`) — the other parts are page URLs, not assets.
- * @param {string} input input
- * @returns {ParsedSource[]} parsed icon-uri
- */
-const parseMsapplicationTask = (input) => {
-	const len = input.length;
-	let pos = 0;
-	while (pos < len) {
-		let sep = input.indexOf(";", pos);
-		if (sep === -1) sep = len;
-		const eq = input.indexOf("=", pos);
-		if (eq !== -1 && eq < sep) {
-			const key = input.slice(pos, eq).trim().toLowerCase();
-			if (key === "icon-uri") {
-				let start = eq + 1;
-				let end = sep;
-				while (start < end && isASCIIWhitespace(input.charCodeAt(start))) {
-					start++;
-				}
-				while (end > start && isASCIIWhitespace(input.charCodeAt(end - 1))) {
-					end--;
-				}
-				if (start === end) return [];
-				return [[input.slice(start, end), start, end]];
-			}
-		}
-		pos = sep + 1;
-	}
-	return [];
-};
-
-/**
- * Extracts `url(...)` references from a CSS value (an SVG presentation
- * attribute). Reuses webpack's CSS lexer so quoting/escaping match the CSS
- * spec; returns the `parseSrc` shape so the shared emit path maps and rewrites
- * the spans. Unquoted `url(path)` rewrites the content span; quoted
- * `url("path")` rewrites the inner string span (quotes preserved).
- * @param {string} input attribute value (a CSS component-value list)
- * @returns {ParsedSource[]} url references
- */
-const parseCssUrls = (input) => {
-	const ts = new TokenStream(input);
-	/** @type {ParsedSource[]} */
-	const result = [];
-	for (;;) {
-		const t = ts.consume();
-		if (t.type === TT_EOF) break;
-		if (t.type === TT_URL) {
-			if (t.contentEnd > t.contentStart) {
-				result.push([
-					input.slice(t.contentStart, t.contentEnd),
-					t.contentStart,
-					t.contentEnd
-				]);
-			}
-		} else if (
-			t.type === TT_FUNCTION &&
-			equalsLowerCase(input.slice(t.start, t.end - 1), "url")
-		) {
-			let s = ts.consume();
-			while (s.type === TT_WHITESPACE) s = ts.consume();
-			if (s.type === TT_STRING) {
-				const quote = input.charCodeAt(s.start);
-				const innerStart = s.start + 1;
-				// Drop the closing quote, unless the string is unterminated at EOF.
-				const innerEnd =
-					input.charCodeAt(s.end - 1) === quote ? s.end - 1 : s.end;
-				if (innerEnd > innerStart) {
-					result.push([
-						input.slice(innerStart, innerEnd),
-						innerStart,
-						innerEnd
-					]);
-				}
-			}
-		}
-	}
-	return result;
-};
 
 /**
  * @param {Map<string, string>} attributes attributes
@@ -1165,7 +962,7 @@ class HtmlParser extends Parser {
 				let attrStart = path.attributeNameStart(typeAttr);
 				if (
 					attrStart > 0 &&
-					isASCIIWhitespace(input.charCodeAt(attrStart - 1))
+					isAsciiWhitespace(input.charCodeAt(attrStart - 1))
 				) {
 					attrStart -= 1;
 				}
@@ -1695,7 +1492,7 @@ class HtmlParser extends Parser {
 									/** @type {number[] | undefined} */
 									let map;
 									if (value.includes("&")) {
-										({ text, map } = decodeEntitiesWithMap(value, true));
+										({ text, map } = decodeEntities(value, true, true));
 										if (!/\S/.test(text)) break;
 									}
 									/** @type {ParsedSource[] | undefined} */
