+++ lib/html/syntax.js
@@ -3775,15 +3775,6 @@ const decodeNumericReference = (code) => {
 	return String.fromCodePoint(code);
 };
 
-// Match one of three forms (each with an optional trailing `;`):
-//   `&#x<hex>` - hex numeric reference (requires the `x`/`X`).
-//   `&#<dec>`  - decimal numeric reference (digits only).
-//   `&<name>`  - named reference (letter followed by alphanumerics).
-// The three alternatives are kept separate so a decimal reference like
-// `&#65b` doesn't greedily eat the trailing `b` as if it were hex.
-const CHARACTER_REFERENCE_REGEXP =
-	/&(?:#[xX][0-9a-fA-F]+|#[0-9]+|[a-zA-Z][a-zA-Z0-9]*);?/g;
-
 /**
  * Decode a single matched character reference.
  * @param {string} match the matched reference text
@@ -3865,6 +3856,8 @@ const decodeOneReference = (match, nextCharCode, isAttribute) => {
 	return match;
 };
 
+/** @typedef {{ text: string, map: number[] | undefined }} DecodedEntitiesWithMap */
+
 /**
  * Decode HTML character references in a string. Handles all numeric
  * references (with WHATWG remap of 0x00, surrogates, out-of-range, and the
@@ -3876,19 +3869,45 @@ const decodeOneReference = (match, nextCharCode, isAttribute) => {
  * trailing `;` whose next character is `=` or ASCII alphanumeric is left
  * undecoded, so e.g. `&amp=foo` stays literal in an attribute value but
  * decodes to `&=foo` in text.
+ *
+ * With `withMap` the result also carries a boundary map from the decoded
+ * string back to raw offsets, so spans computed on the decoded text (e.g.
+ * srcset candidate URLs) can be translated to source ranges. `map[i]` is the
+ * raw offset of decoded boundary `i` (`0..text.length`); boundaries inside a
+ * reference's decoded text map to the reference start. `map` is `undefined`
+ * when nothing was decoded (then `text === str`).
+ * @overload
  * @param {string} str the raw string from the token slice
  * @param {boolean=} isAttribute true if `str` came from an attribute value
  * @returns {string} decoded string
  */
-const decodeEntities = (str, isAttribute) => {
-	// Hand-rolled scan of the `CHARACTER_REFERENCE_REGEXP` shape
-	// (`&#x<hex>+` / `&#<dec>+` / `&<alpha><alnum>*`, optional `;`): entity-dense
-	// text pays no regex machinery or per-match object, and a string whose
-	// references all stay literal is returned unchanged without rebuilding.
+/**
+ * @overload
+ * @param {string} str the raw string from the token slice
+ * @param {boolean | undefined} isAttribute true if `str` came from an attribute value
+ * @param {true} withMap also build the decoded-to-raw boundary map
+ * @returns {DecodedEntitiesWithMap} decoded text and offset map
+ */
+/**
+ * @param {string} str the raw string from the token slice
+ * @param {boolean=} isAttribute true if `str` came from an attribute value
+ * @param {boolean=} withMap also build the decoded-to-raw boundary map
+ * @returns {string | DecodedEntitiesWithMap} decoded string (with map when `withMap`)
+ */
+const decodeEntities = (str, isAttribute, withMap) => {
+	// Hand-rolled scan of one of three reference forms, each with an optional
+	// trailing `;` — `&#x<hex>+` / `&#<dec>+` / `&<alpha><alnum>*` (kept separate
+	// so `&#65b` doesn't eat the `b` as hex): entity-dense text pays no regex
+	// machinery or per-match object, and a string whose references all stay
+	// literal is returned unchanged without rebuilding. Map bookkeeping is
+	// gated on `withMap` at the rare per-decoded-reference points, so the
+	// dominant no-map mode pays nothing for the unified implementation.
 	let amp = str.indexOf("&");
-	if (amp === -1) return str;
+	if (amp === -1) return withMap ? { text: str, map: undefined } : str;
 	let out = "";
 	let last = 0;
+	/** @type {number[] | undefined} */
+	let map;
 	do {
 		// `end` stays 0 when the `&` doesn't start a well-formed reference.
 		let end = 0;
@@ -3919,6 +3938,7 @@ const decodeEntities = (str, isAttribute) => {
 				isAttribute
 			);
 			if (decoded !== match) {
+				if (withMap) map = _mapReference(map, last, amp, decoded.length);
 				out += str.slice(last, amp);
 				out += decoded;
 				last = end;
@@ -3928,51 +3948,40 @@ const decodeEntities = (str, isAttribute) => {
 			amp = str.indexOf("&", amp + 1);
 		}
 	} while (amp !== -1);
-	return last === 0 ? str : out + str.slice(last);
+	if (last === 0) return withMap ? { text: str, map: undefined } : str;
+	const text = out + str.slice(last);
+	return withMap
+		? _finishMap(text, /** @type {number[]} */ (map), last, str.length)
+		: text;
+};
+
+// Map-mode bookkeeping for `decodeEntities`, kept out of the scanner body so
+// the dominant no-map mode stays small enough to inline.
+/**
+ * @param {number[] | undefined} map boundary map so far
+ * @param {number} last raw offset the literal run starts at
+ * @param {number} amp raw offset of the decoded reference
+ * @param {number} decodedLength length of the decoded text
+ * @returns {number[]} the map, with the literal run's and reference's boundaries appended
+ */
+const _mapReference = (map, last, amp, decodedLength) => {
+	if (map === undefined) map = [];
+	for (let r = last; r < amp; r++) map.push(r);
+	for (let i = 0; i < decodedLength; i++) map.push(amp);
+	return map;
 };
 
 /**
- * Like `decodeEntities`, but also returns a boundary map from the
- * decoded string back to raw offsets, so spans computed on the decoded
- * text (e.g. srcset candidate URLs) can be translated to source ranges.
- * `map[i]` is the raw offset of decoded boundary `i` (`0..text.length`);
- * boundaries inside a reference's decoded text map to the reference start.
- * `map` is `undefined` when nothing was decoded (then `text === str`).
- * @param {string} str the raw string from the token slice
- * @param {boolean=} isAttribute true if `str` came from an attribute value
- * @returns {{ text: string, map: number[] | undefined }} decoded text and offset map
+ * @param {string} text decoded text
+ * @param {number[]} map boundary map so far
+ * @param {number} last raw offset of the trailing literal run
+ * @param {number} rawLength raw string length
+ * @returns {DecodedEntitiesWithMap} the finished result
  */
-const decodeEntitiesWithMap = (str, isAttribute) => {
-	/** @type {number[] | undefined} */
-	let map;
-	if (str.includes("&")) {
-		CHARACTER_REFERENCE_REGEXP.lastIndex = 0;
-		let text = "";
-		let last = 0;
-		/** @type {RegExpExecArray | null} */
-		let m;
-		while ((m = CHARACTER_REFERENCE_REGEXP.exec(str)) !== null) {
-			const match = m[0];
-			const decoded = decodeOneReference(
-				match,
-				str.charCodeAt(m.index + match.length),
-				isAttribute
-			);
-			if (decoded === match) continue;
-			if (map === undefined) map = [];
-			for (let r = last; r < m.index; r++) map.push(r);
-			text += str.slice(last, m.index);
-			for (let i = 0; i < decoded.length; i++) map.push(m.index);
-			text += decoded;
-			last = m.index + match.length;
-		}
-		if (map !== undefined) {
-			for (let r = last; r < str.length; r++) map.push(r);
-			map.push(str.length);
-			return { text: text + str.slice(last), map };
-		}
-	}
-	return { text: str, map };
+const _finishMap = (text, map, last, rawLength) => {
+	for (let r = last; r < rawLength; r++) map.push(r);
+	map.push(rawLength);
+	return { text, map };
 };
 
 /**
@@ -8575,6 +8584,198 @@ const parseSrcset = (input) => {
 	}
 };
 
+// The spec's "URL-potentially-surrounded-by-spaces" cleanup: C0 controls,
+// DEL..C1 and U+00A0 are stripped from the value after the whitespace trim.
+// eslint-disable-next-line no-control-regex
+const IGNORE_CHARS_REGEXP = /[\u0000-\u001F\u007F-\u009F\u00A0]/g;
+
+/**
+ * Parse a `src`-like URL attribute value: trim ASCII whitespace / U+00A0 from
+ * both ends (offsets preserved for source rewriting), then strip ignorable
+ * control characters. Throws when nothing remains.
+ * @param {string} input attribute value
+ * @returns {ParsedSource[]} parsed src
+ */
+const parseSrc = (input) => {
+	const len = input.length;
+	if (len === 0) throw new Error("Must be non-empty");
+
+	let start = 0;
+	let end = len;
+
+	while (start < end) {
+		const code = input.charCodeAt(start);
+		if (code > 32 && code !== 160) break;
+		start++;
+	}
+
+	if (start === end) throw new Error("Must be non-empty");
+
+	while (end > start) {
+		const code = input.charCodeAt(end - 1);
+		if (code > 32 && code !== 160) break;
+		end--;
+	}
+
+	let value = input.slice(start, end);
+
+	if (IGNORE_CHARS_REGEXP.test(value)) {
+		value = value.replace(IGNORE_CHARS_REGEXP, "");
+		if (value.length === 0) throw new Error("Must be non-empty");
+	}
+
+	return [[value, start, end]];
+};
+
+/**
+ * Extracts the `icon-uri` value of an `msapplication-task` meta content
+ * (`name=…;action-uri=…;icon-uri=…`) — the other parts are page URLs, not assets.
+ * @param {string} input input
+ * @returns {ParsedSource[]} parsed icon-uri
+ */
+const parseMsapplicationTask = (input) => {
+	const len = input.length;
+	let pos = 0;
+	while (pos < len) {
+		let sep = input.indexOf(";", pos);
+		if (sep === -1) sep = len;
+		const eq = input.indexOf("=", pos);
+		if (eq !== -1 && eq < sep) {
+			const key = input.slice(pos, eq).trim().toLowerCase();
+			if (key === "icon-uri") {
+				let start = eq + 1;
+				let end = sep;
+				while (start < end && isSpace(input.charCodeAt(start))) {
+					start++;
+				}
+				while (end > start && isSpace(input.charCodeAt(end - 1))) {
+					end--;
+				}
+				if (start === end) return [];
+				return [[input.slice(start, end), start, end]];
+			}
+		}
+		pos = sep + 1;
+	}
+	return [];
+};
+
+// CSS syntax is loaded on first `parseCssUrls` call, not at module load —
+// `HtmlGenerator` deliberately defers it the same way, and most documents
+// never carry a URL-bearing SVG presentation attribute.
+/** @type {typeof import("../css/syntax") | undefined} */
+let _cssSyntax;
+
+/**
+ * Extracts `url(...)` references from a CSS value (an SVG presentation
+ * attribute). Reuses webpack's CSS lexer so quoting/escaping match the CSS
+ * spec; returns the `parseSrc` shape so the shared emit path maps and rewrites
+ * the spans. Unquoted `url(path)` rewrites the content span; quoted
+ * `url("path")` rewrites the inner string span (quotes preserved).
+ * @param {string} input attribute value (a CSS component-value list)
+ * @returns {ParsedSource[]} url references
+ */
+const parseCssUrls = (input) => {
+	const {
+		TT_EOF,
+		TT_FUNCTION,
+		TT_STRING,
+		TT_URL,
+		TT_WHITESPACE,
+		TokenStream,
+		equalsLowerCase
+	} = _cssSyntax || (_cssSyntax = require("../css/syntax"));
+	const ts = new TokenStream(input);
+	/** @type {ParsedSource[]} */
+	const result = [];
+	for (;;) {
+		const t = ts.consume();
+		if (t.type === TT_EOF) break;
+		if (t.type === TT_URL) {
+			if (t.contentEnd > t.contentStart) {
+				result.push([
+					input.slice(t.contentStart, t.contentEnd),
+					t.contentStart,
+					t.contentEnd
+				]);
+			}
+		} else if (
+			t.type === TT_FUNCTION &&
+			equalsLowerCase(input.slice(t.start, t.end - 1), "url")
+		) {
+			let s = ts.consume();
+			while (s.type === TT_WHITESPACE) s = ts.consume();
+			if (s.type === TT_STRING) {
+				const quote = input.charCodeAt(s.start);
+				const innerStart = s.start + 1;
+				// Drop the closing quote, unless the string is unterminated at EOF.
+				const innerEnd =
+					input.charCodeAt(s.end - 1) === quote ? s.end - 1 : s.end;
+				if (innerEnd > innerStart) {
+					result.push([
+						input.slice(innerStart, innerEnd),
+						innerStart,
+						innerEnd
+					]);
+				}
+			}
+		}
+	}
+	return result;
+};
+
+/**
+ * Byte-exact source span of an attribute including the single leading
+ * whitespace (` name`, ` name=value`, ` name="value"`), mirroring the
+ * tokenizer's end-of-attribute rule (`valueEnd + 1` past the closing quote).
+ * @param {HtmlPath} path walk path (used only for attribute-ref reads)
+ * @param {string} source HTML source
+ * @param {HtmlAttributeRef} attr attribute ref
+ * @returns {string} the attribute's source slice
+ */
+const attrSourceSpan = (path, source, attr) => {
+	const valueStart = path.attributeValueStart(attr);
+	const valueEnd = path.attributeValueEnd(attr);
+	const end =
+		valueStart === -1
+			? path.attributeNameEnd(attr)
+			: source.charCodeAt(valueStart - 1) === CC_QUOTATION_MARK ||
+				  source.charCodeAt(valueStart - 1) === CC_APOSTROPHE
+				? valueEnd + 1
+				: valueEnd;
+	return source.slice(path.attributeNameStart(attr) - 1, end);
+};
+
+/**
+ * Source span of a tag's `integrity` attribute (offsets relative to
+ * `elementStart`, leading whitespace and quotes included), so the template can
+ * drop the content-specific author value without re-parsing. Null when absent.
+ * @param {HtmlPath} path walk path (for attribute reads)
+ * @param {string} source HTML source
+ * @param {number} elementStart offset of the tag's opening `<`
+ * @returns {[number, number] | null} the span, or null when there is no `integrity`
+ */
+const captureIntegrityRange = (path, source, elementStart) => {
+	const integrityAttr = path.findAttribute("integrity");
+	if (integrityAttr === 0 || path.attributeNameStart(integrityAttr) < 0) {
+		return null;
+	}
+	let rs = path.attributeNameStart(integrityAttr);
+	while (rs > 0 && isSpace(source.charCodeAt(rs - 1))) {
+		rs--;
+	}
+	let re;
+	const ivs = path.attributeValueStart(integrityAttr);
+	if (ivs === -1) {
+		re = path.attributeNameEnd(integrityAttr);
+	} else {
+		const q = source.charCodeAt(ivs - 1);
+		const ive = path.attributeValueEnd(integrityAttr);
+		re = q === CC_QUOTATION_MARK || q === CC_APOSTROPHE ? ive + 1 : ive;
+	}
+	return [rs - elementStart, re - elementStart];
+};
+
 // Babel's `path.skip()`, children-only: set by `A.skipChildren()` during an
 // `enter` dispatch, consumed by the walk.
 let _walkSkip = false;
@@ -8848,14 +9049,21 @@ module.exports.NodeType = NodeType;
 module.exports.QUOTE_DOUBLE = QUOTE_DOUBLE;
 module.exports.QUOTE_NONE = QUOTE_NONE;
 module.exports.QUOTE_SINGLE = QUOTE_SINGLE;
+// WHATWG "ASCII whitespace" (tab / LF / FF / CR / space) — the tokenizer's
+// whitespace class, exported under the spec's name.
 // Exposed so HtmlParser can map user-configured (lowercased) tag names to
 // the adjusted camelCase names the AST carries for foreign content.
 module.exports.SVG_TAG_ADJUST = SVG_TAG_ADJUST;
 module.exports.SourceProcessor = SourceProcessor;
+module.exports.attrSourceSpan = attrSourceSpan;
+module.exports.captureIntegrityRange = captureIntegrityRange;
 module.exports.decodeEntities = decodeEntities;
-module.exports.decodeEntitiesWithMap = decodeEntitiesWithMap;
 module.exports.escapeAttribute = escapeAttribute;
 module.exports.escapeText = escapeText;
+module.exports.isAsciiWhitespace = isSpace;
+module.exports.parseCssUrls = parseCssUrls;
 module.exports.parseHtml = parseHtml;
+module.exports.parseMsapplicationTask = parseMsapplicationTask;
+module.exports.parseSrc = parseSrc;
 module.exports.parseSrcset = parseSrcset;
 module.exports.tokenize = tokenize;
